Move useParamStore to common
