+++ airflow/ui/src/components/FlexibleForm/HiddenInput.tsx
@@ -19,7 +19,7 @@
 import { VisuallyHidden } from "@chakra-ui/react";
 
 import type { FlexibleFormElementProps } from ".";
-import { paramPlaceholder, useParamStore } from "../TriggerDag/useParamStore";
+import { paramPlaceholder, useParamStore } from "../../queries/useParamStore";
 
 /** Render a "const" field where user can not change data as hidden */
 export const HiddenInput = ({ name }: FlexibleFormElementProps) => {
